Fix all PHPStan errors (#265)
Level 8 analysis of game/ and wwwroot/ now passes with zero errors
(previously 602+ errors). All fixes are static-analysis-only:

- Add type declarations / PHPDoc types to legacy functions and properties
- Add null/false guards and array shape annotations where DB fetches
  can return null/false
- Cast ambiguous values (int/string/float) at call sites
- Fix real bugs found by the analyzer:
  * unquoted array keys in botapi.php ([player_id])
  * filectime() typo in flottenversand_ajax.php ( -> )
  * each() removed in PHP 8 (bbcode.php)
  * relative include paths in admin_db.php/admin_queue.php (__DIR__)
  * missing return in AddChangeEmailEvent()
- Remove provably dead code and constant conditions (no behavior change)
- game/core/bbcode.php is no longer excluded from analysis
- phpstan.neon: ignore config.php requires (file generated at install time)
- messages.php: keep intentional method-selection structure, mark the
  dormant branches with @phpstan-ignore comments

Verified: phpstan analyze -> [OK] No errors; phpunit -> OK (167 tests,
932 assertions); phpstan_report_latest.txt updated.

+++ game/pages/trader.php
@@ -41,9 +41,9 @@ class Trader extends Page {
                     $value_2 = 0;
                     $value_3 = 0;
 
-                    if (key_exists('1_value', $_POST)) $value_1 = abs (str_replace ( ".", "", $_POST['1_value'] ));
-                    if (key_exists('2_value', $_POST)) $value_2 = abs (str_replace ( ".", "", $_POST['2_value'] ));
-                    if (key_exists('3_value', $_POST)) $value_3 = abs (str_replace ( ".", "", $_POST['3_value'] ));
+                    if (key_exists('1_value', $_POST)) $value_1 = abs ( (int) str_replace ( ".", "", $_POST['1_value'] ) );
+                    if (key_exists('2_value', $_POST)) $value_2 = abs ( (int) str_replace ( ".", "", $_POST['2_value'] ) );
+                    if (key_exists('3_value', $_POST)) $value_3 = abs ( (int) str_replace ( ".", "", $_POST['3_value'] ) );
 
                     if ( $GlobalUser['trader'] == 1) {
                         $crys = floor ( $aktplanet[GID_RC_CRYSTAL] + $value_2 );
@@ -114,9 +114,9 @@ class Trader extends Page {
         // Prepare display data
         if ( $GlobalUser['trader'] > 0 ) {
             $this->offer_id = $GlobalUser['trader'];
-            if ( $this->offer_id == 1) $this->amount = floor ($aktplanet[GID_RC_METAL]);
-            else if ( $this->offer_id == 2) $this->amount = floor ($aktplanet[GID_RC_CRYSTAL]);
-            else if ( $this->offer_id == 3) $this->amount = floor ($aktplanet[GID_RC_DEUTERIUM]);
+            if ( $this->offer_id == 1) $this->amount = (int) floor ($aktplanet[GID_RC_METAL]);
+            else if ( $this->offer_id == 2) $this->amount = (int) floor ($aktplanet[GID_RC_CRYSTAL]);
+            else if ( $this->offer_id == 3) $this->amount = (int) floor ($aktplanet[GID_RC_DEUTERIUM]);
             $mmax = max (0, $aktplanet['max'.GID_RC_METAL] - $aktplanet[GID_RC_METAL] );
             $kmax = max (0, $aktplanet['max'.GID_RC_CRYSTAL] - $aktplanet[GID_RC_CRYSTAL] );
             $dmax = max (0, $aktplanet['max'.GID_RC_DEUTERIUM] - $aktplanet[GID_RC_DEUTERIUM] );
